+++ CHANGELOG.md
@@ -4,12 +4,13 @@
 ## ✨ New Features
 - Adds `Item3D.intersects(item: Item3D, precision: "low" | "medium")` to test for intersections of Item3D objects and their associated meshes.
 - Moves clongin logic to `Item3D`, adding `Item3D.clone()` method for easier cloning of 3D items.
+- Adds `remove()` method to `Item3D`
 
 ## ⚙️ Other
-None
+- Addes e2e tests for intersections
 
 ## 🐞 Bug Fixes
-None
+- Fixes `Layer3D.removeMesh` logic
 
 ### 3.1.0